Refactor: useLocalCart 내 useEffect 추가
